Move tt files to riscv/tt subdirectory
