+++ gcc/Makefile.in
@@ -1410,15 +1410,6 @@ OBJS = \
 	gimple-range-fold.o \
 	gimple-range-gori.o \
 	gimple-range-trace.o \
-	gimple-rvtt-attrib.o \
-	gimple-rvtt-cc.o \
-	gimple-rvtt-combine.o \
-	gimple-rvtt-expand.o \
-	gimple-rvtt-live.o \
-	gimple-rvtt-move.o \
-	gimple-rvtt-nonimm-expand.o \
-	gimple-rvtt-nonimm-tag.o \
-	gimple-rvtt-warn.o \
 	gimple-ssa-backprop.o \
 	gimple-ssa-evrp.o \
 	gimple-ssa-evrp-analyze.o \
@@ -1573,11 +1564,6 @@ OBJS = \
 	rtl-ssa/changes.o \
 	rtl-ssa/functions.o \
 	rtl-ssa/insns.o \
-	rtl-rvtt-hll.o \
-	rtl-rvtt-nonimm.o \
-	rtl-rvtt-replay.o \
-	rtl-rvtt-rmext.o \
-	rtl-rvtt-schedule.o \
 	rtl-tests.o \
 	rtl.o \
 	rtlhash.o \
