+++ gcc/config.gcc
@@ -517,7 +517,7 @@ pru-*-*)
 	;;
 riscv*)
 	cpu_type=riscv
-	extra_objs="riscv-builtins.o riscv-c.o riscv-sr.o riscv-shorten-memrefs.o rvtt.o rvtt-gs.o rvtt-wh.o rvtt-bh.o"
+	extra_objs="riscv-builtins.o riscv-c.o riscv-sr.o riscv-shorten-memrefs.o"
 	d_target_objs="riscv-d.o"
 	;;
 rs6000*-*-*)
@@ -2477,6 +2477,7 @@ riscv*-*-elf* | riscv*-*-rtems*)
 	  fi
 	esac
 	tmake_file="${tmake_file} riscv/t-riscv"
+	tmake_file="${tmake_file} riscv/tt/t-riscv-tt"
 	gnu_ld=yes
 	gas=yes
 	case $target in
