+++ gcc/config/riscv/tt/gimple-rvtt-live.cc
@@ -65,7 +65,7 @@ along with GCC; see the file COPYING3.  If not see
 #include <vector>
 #include <unordered_map>
 #include <tuple>
-#include "config/riscv/rvtt.h"
+#include "rvtt.h"
 
 #define DUMP(...) //fprintf(stderr, __VA_ARGS__)
