+++ gcc/config/riscv/tt/gimple-rvtt-move.cc
@@ -67,7 +67,7 @@ along with GCC; see the file COPYING3.  If not see
 #include <string>
 #include <iostream>
 #include <tuple>
-#include "config/riscv/rvtt.h"
+#include "rvtt.h"
 
 #define DUMP(...) //fprintf(stderr, __VA_ARGS__)
