+++ gcc/config/riscv/tt/rtl-rvtt-rmext.cc
@@ -38,7 +38,7 @@
 #include "function-abi.h"
 #include <vector>
 #include <unordered_map>
-#include "config/riscv/rvtt.h"
+#include "rvtt.h"
 
 #define DUMP(...) //fprintf(stderr, __VA_ARGS__)
