+++ gcc/config/riscv/tt/rvtt-gs.cc
