+++ gcc/config/riscv/tt/rvtt-gs.md
@@ -907,4 +907,4 @@
   "TARGET_RVTT_GS"
   "TTNOP")
 
-(include "rvtt-peephole-gs.md")
+(include "tt/rvtt-peephole-gs.md")
