+++ gcc/config/riscv/tt/rvtt-peephole-gs.md
