+++ gcc/config/riscv/tt/rvtt-peephole-wh.md
