+++ gcc/config/riscv/tt/rvtt-protos.h
