+++ gcc/config/riscv/tt/rvtt-wh.cc
