+++ gcc/config/riscv/tt/rvtt.md
@@ -19,8 +19,8 @@
 ;; along with GCC; see the file COPYING3.  If not see
 ;; <http://www.gnu.org/licenses/>.
 
-(include "rvtt-predicates.md")
-(include "rvtt-tune.md")
+(include "tt/rvtt-predicates.md")
+(include "tt/rvtt-tune.md")
 
 ; & in spec means early clobber, written before inputs are used, cannot reuse input reg
